Spincam has been moved to its own file, cleaning up Kitty.cs a little.
Teamdeathless update.. -- TO DO: Orb carrier shouldn't be able to get summoned or use shadow kitty
