+++ Source/Events/Commands/COMMAND REVAMP/InitCommands.cs
@@ -422,8 +422,8 @@ public static class InitCommands
             action: (player, args) =>
             {
                 float speed = args[0] != "" ? float.Parse(args[0]) : 0;
-                Globals.ALL_KITTIES[player].ToggleSpinCam(speed);
-                player.DisplayTextTo(Colors.COLOR_GOLD + "SpinCam: " + (Globals.ALL_KITTIES[player].IsSpinCamActive() ? "On" : "Off"));
+                Globals.ALL_KITTIES[player].SpinCam.ToggleSpinCam(speed);
+                player.DisplayTextTo(Colors.COLOR_GOLD + "SpinCam: " + (Globals.ALL_KITTIES[player].SpinCam.IsSpinCamActive() ? "On" : "Off"));
             }
         );
